no force_tenant_isolation in DefaultDomainTestJSON
No need for dynamic credentials to test showing the default domain.

By setting this value to False, consumers with an immutable user source
can execute this test.

Depends-On: I83a9b8af775580d36a1141be55e9c1cc283a75b6
Partial-Bug: #1714277
